+++ template.yaml
@@ -7,19 +7,30 @@ Description: >
 
 Globals:
   Function:
-    Timeout: 3
+    Timeout: 30
     MemorySize: 128
+
+Parameters:
+  S3BucketName:
+    Type: String
+    Default: "sample-app-bucket"
+  MessageTableName:
+    Type: String
+    Default: "Message"
+  MemberTableName:
+    Type: String
+    Default: "Member"
 
 Resources:
   # S3: test-data bucket
-  TestDataBucket:
+  SampleAppBucket:
     Type: AWS::S3::Bucket
     Properties:
       AccessControl: PublicReadWrite
-      BucketName: test-data
+      BucketName: !Ref S3BucketName
 
-  # DynamoDB: UserMessage table
-  UserMessageTable:
+  # DynamoDB: Message table
+  MessageTable:
     Type: AWS::DynamoDB::Table
     Properties:
       AttributeDefinitions:
@@ -28,7 +39,7 @@ Resources:
       KeySchema:
         - AttributeName: seq
           KeyType: HASH
-      TableName: UserMessage
+      TableName: !Ref MessageTableName
       ProvisionedThroughput:
         ReadCapacityUnits: 5
         WriteCapacityUnits: 5
@@ -43,7 +54,7 @@ Resources:
       KeySchema:
         - AttributeName: seq
           KeyType: HASH
-      TableName: Member
+      TableName: !Ref MemberTableName
       ProvisionedThroughput:
         ReadCapacityUnits: 5
         WriteCapacityUnits: 5
@@ -72,29 +83,29 @@ Resources:
       Architectures:
         - x86_64
 
-  # Lambda: SimpleDynamodbFunction
-  SimpleDynamodbFunction:
+  # Lambda: AddMessageFunction
+  AddMessageFunction:
     Type: AWS::Serverless::Function
     Properties:
       CodeUri: function/
-      Handler: simple_dynamodb.lambda_handler
+      Handler: add_message.lambda_handler
       Runtime: python3.9
       Role: arn:aws:iam::000000000000:role/lambda-role
       Architectures:
         - x86_64
 
-  # Lambda: WorkWithOtherFunction
-  WorkWithOtherFunction:
+  # Lambda: UpdateMemberFileFunction
+  UpdateMemberFileFunction:
     Type: AWS::Serverless::Function
     Properties:
       CodeUri: function/
-      Handler: work_with_other.lambda_handler
+      Handler: update_member_file.lambda_handler
       Runtime: python3.9
       Role: arn:aws:iam::000000000000:role/lambda-role
       Architectures:
         - x86_64
       Events:
-        WorkWithOther:
+        MemberTableUpdate:
           Type: DynamoDB
           Properties:
             Stream: !GetAtt MemberTable.StreamArn
